Run app as non-root
In docker the user namespace is shared between the host and docker
containers, so root in a container is root on the host.

In the event of the app being compromised running the attacker will only
have access to that user, rather than to the root of the host and all
the containers on the machine. Much better. :)

I should have done this first time round. Oops.